Move node visitor class to expr.
Done by:
  * Running `rg 'smt_util/node_visitor' -l | xargs sed -i 's/smt_util\/node_visitor/expr\/node_visitor/'` to change the #includes
  * Moving the file
  * Changing src/expr/CMakeLists.txt and src/CMakeLists.txt
