+++ src/preprocessing/passes/bv_to_bool.cpp
@@ -23,7 +23,7 @@
 
 #include "expr/node.h"
 #include "smt/smt_statistics_registry.h"
-#include "smt_util/node_visitor.h"
+#include "expr/node_visitor.h"
 #include "theory/rewriter.h"
 #include "theory/theory.h"
