+++ src/printer/ast/ast_printer.cpp
@@ -25,7 +25,7 @@
 #include "options/language.h" // for LANG_AST
 #include "printer/dagification_visitor.h"
 #include "smt/command.h"
-#include "smt_util/node_visitor.h"
+#include "expr/node_visitor.h"
 #include "theory/substitutions.h"
 
 using namespace std;
